Move unit tests to correct file
